issue 컴포넌트 랜더링 코드 분리
